pyproject.toml
[project]
name = "send-mail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "secure-smtplib>=0.1.1",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.38.0",
]
